+++ system-configuration/examples/watch_dns.rs
@@ -1,14 +1,8 @@
-use core_foundation::{
-    array::CFArray,
-    base::{CFType, TCFType, ToVoid},
-    dictionary::CFDictionary,
-    propertylist::CFPropertyList,
-    runloop::{kCFRunLoopCommonModes, CFRunLoop},
-    string::CFString,
-};
+use objc2_core_foundation::{kCFRunLoopCommonModes, CFRetained, CFRunLoop, CFType};
 use system_configuration::{
+    core_foundation::{CFArray, CFDictionary, CFString},
     dynamic_store::{SCDynamicStore, SCDynamicStoreBuilder, SCDynamicStoreCallBackContext},
-    sys::schema_definitions::kSCPropNetDNSServerAddresses,
+    sys::kSCPropNetDNSServerAddresses,
 };
 
 // This example will watch the dynamic store for changes to any DNS setting. As soon as a change
@@ -25,9 +19,10 @@ fn main() {
         .build()
         .expect("Unable to create DynamicStore");
 
-    let watch_keys: CFArray<CFString> = CFArray::from_CFTypes(&[]);
-    let watch_patterns =
-        CFArray::from_CFTypes(&[CFString::from("(State|Setup):/Network/Service/.*/DNS")]);
+    let watch_keys = CFArray::<CFString>::from_objects(&[]);
+    let watch_patterns = CFArray::from_retained_objects(&[CFString::from_str(
+        "(State|Setup):/Network/Service/.*/DNS",
+    )]);
 
     if store.set_notification_keys(&watch_keys, &watch_patterns) {
         println!("Registered for notifications");
@@ -38,11 +33,11 @@ fn main() {
     let run_loop_source = store
         .create_run_loop_source()
         .expect("Unable to create run loop source");
-    let run_loop = CFRunLoop::get_current();
-    run_loop.add_source(&run_loop_source, unsafe { kCFRunLoopCommonModes });
+    let run_loop = CFRunLoop::current().unwrap();
+    run_loop.add_source(Some(&run_loop_source), unsafe { kCFRunLoopCommonModes });
 
     println!("Entering run loop");
-    CFRunLoop::run_current();
+    CFRunLoop::run();
 }
 
 /// This struct acts as a user provided context/payload to each notification callback.
@@ -53,12 +48,12 @@ struct Context {
 }
 
 #[allow(clippy::needless_pass_by_value)]
-fn my_callback(store: SCDynamicStore, changed_keys: CFArray<CFString>, context: &mut Context) {
+fn my_callback(store: &SCDynamicStore, changed_keys: &CFArray<CFString>, context: &mut Context) {
     context.call_count += 1;
     println!("Callback call count: {}", context.call_count);
 
     for key in changed_keys.iter() {
-        if let Some(addresses) = get_dns(&store, key.clone()) {
+        if let Some(addresses) = get_dns(store, &key) {
             println!("{} changed DNS to {:?}", *key, addresses);
         } else {
             println!("{} removed DNS", *key);
@@ -66,18 +61,18 @@ fn my_callback(store: SCDynamicStore, changed_keys: CFArray<CFString>, context: 
     }
 }
 
-fn get_dns(store: &SCDynamicStore, path: CFString) -> Option<Vec<String>> {
+fn get_dns(store: &SCDynamicStore, path: &CFString) -> Option<Vec<String>> {
     let dns_settings = store
-        .get(path)
-        .and_then(CFPropertyList::downcast_into::<CFDictionary>)?;
+        .get(path.to_string())
+        .and_then(|ty| CFRetained::downcast::<CFDictionary>(ty).ok())?;
+    let dns_settings = unsafe { dns_settings.cast_unchecked::<CFString, CFType>() };
     let address_array = dns_settings
-        .find(unsafe { kSCPropNetDNSServerAddresses }.to_void())
-        .map(|ptr| unsafe { CFType::wrap_under_get_rule(*ptr) })
-        .and_then(CFType::downcast_into::<CFArray>)?;
+        .get(unsafe { kSCPropNetDNSServerAddresses })
+        .and_then(|ty| CFRetained::downcast::<CFArray>(ty).ok())?;
+    let address_array = unsafe { address_array.cast_unchecked::<CFType>() };
     let mut result = Vec::with_capacity(address_array.len() as usize);
-    for address_ptr in &address_array {
-        let address =
-            unsafe { CFType::wrap_under_get_rule(*address_ptr) }.downcast_into::<CFString>()?;
+    for address in address_array {
+        let address = address.downcast::<CFString>().ok()?;
         result.push(address.to_string())
     }
     Some(result)